services/job-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
aiokafka==0.10.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
